Rough splitting into files
